re-factor to use pandas_flavor
