--- .stickler.yml
@@ -1,6 +1,0 @@
-linters:
-  flake8:
-    max-line-length: 120
-    max-complexity: 10
-files:
-  ignore: ['*.ipynb', 'conf.py']
