+++ ctd/extras.py
@@ -1,0 +1,350 @@
+import matplotlib.pyplot as plt
+import mpl_toolkits.axisartist as AA
+import numpy as np
+import numpy.ma as ma
+from mpl_toolkits.axes_grid1 import host_subplot
+from pandas import Series
+
+
+def _extrap1d(interpolator):
+    """
+    http://stackoverflow.com/questions/2745329/
+    How to make scipy.interpolate return an extrapolated result beyond the
+    input range.
+
+    """
+    xs, ys = interpolator.x, interpolator.y
+
+    def pointwise(x):
+        if x < xs[0]:
+            return ys[0] + (x - xs[0]) * (ys[1] - ys[0]) / (xs[1] - xs[0])
+        elif x > xs[-1]:
+            return ys[-1] + (x - xs[-1]) * (ys[-1] - ys[-2]) / (
+                xs[-1] - xs[-2]
+            )
+        else:
+            return interpolator(x)
+
+    def ufunclike(xs):
+        return np.array(list(map(pointwise, np.array(xs))))
+
+    return ufunclike
+
+
+def get_maxdepth(self):
+    valid_last_depth = self.apply(Series.notnull).values.T
+    return np.float_(self.index.values * valid_last_depth).max(axis=1)
+
+
+def extrap_sec(data, dist, depth, w1=1.0, w2=0):
+    """
+    Extrapolates `data` to zones where the shallow stations are shadowed by
+    the deep stations.  The shadow region usually cannot be extrapolates via
+    linear interpolation.
+
+    The extrapolation is applied using the gradients of the `data` at a certain
+    level.
+
+    Parameters
+    ----------
+    data : array_like
+          Data to be extrapolated
+    dist : array_like
+           Stations distance
+    fd : float
+         Decay factor [0-1]
+
+
+    Returns
+    -------
+    Sec_extrap : array_like
+                 Extrapolated variable
+
+    """
+    from scipy.interpolate import interp1d
+
+    new_data1 = []
+    for row in data:
+        mask = ~np.isnan(row)
+        if mask.any():
+            y = row[mask]
+            if y.size == 1:
+                row = np.repeat(y, len(mask))
+            else:
+                x = dist[mask]
+                f_i = interp1d(x, y)
+                f_x = _extrap1d(f_i)
+                row = f_x(dist)
+        new_data1.append(row)
+
+    new_data2 = []
+    for col in data.T:
+        mask = ~np.isnan(col)
+        if mask.any():
+            y = col[mask]
+            if y.size == 1:
+                col = np.repeat(y, len(mask))
+            else:
+                z = depth[mask]
+                f_i = interp1d(z, y)
+                f_z = _extrap1d(f_i)
+                col = f_z(depth)
+        new_data2.append(col)
+
+    new_data = np.array(new_data1) * w1 + np.array(new_data2).T * w2
+    return new_data
+
+
+def gen_topomask(h, lon, lat, dx=1.0, kind="linear", plot=False):
+    """
+    Generates a topography mask from an oceanographic transect taking the
+    deepest CTD scan as the depth of each station.
+
+    Inputs
+    ------
+    h : array
+        Pressure of the deepest CTD scan for each station [dbar].
+    lons : array
+           Longitude of each station [decimal degrees east].
+    lat : Latitude of each station. [decimal degrees north].
+    dx : float
+         Horizontal resolution of the output arrays [km].
+    kind : string, optional
+           Type of the interpolation to be performed.
+           See scipy.interpolate.interp1d documentation for details.
+    plot : bool
+           Whether to plot mask for visualization.
+
+    Outputs
+    -------
+    xm : array
+         Horizontal distances [km].
+    hm : array
+         Local depth [m].
+
+    Author
+    ------
+    André Palóczy Filho (paloczy@gmail.com) --  October/2012
+
+    """
+
+    import gsw
+    from scipy.interpolate import interp1d
+
+    h, lon, lat = list(map(np.asanyarray, (h, lon, lat)))
+    # Distance in km.
+    x = np.append(0, np.cumsum(gsw.distance(lon, lat)[0] / 1e3))
+    h = -gsw.z_from_p(h, lat.mean())
+    Ih = interp1d(x, h, kind=kind, bounds_error=False, fill_value=h[-1])
+    xm = np.arange(0, x.max() + dx, dx)
+    hm = Ih(xm)
+
+    return xm, hm
+
+
+def plot_vars(self, variables=None, **kwds):
+    """
+    Plot CTD temperature and salinity.
+    """
+    fig = plt.figure(figsize=(8, 10))
+    ax0 = host_subplot(111, axes_class=AA.Axes)
+    ax1 = ax0.twiny()
+
+    # Axis location.
+    host_new_axis = ax0.get_grid_helper().new_fixed_axis
+    ax0.axis["bottom"] = host_new_axis(loc="top", axes=ax0, offset=(0, 0))
+    par_new_axis = ax1.get_grid_helper().new_fixed_axis
+    ax1.axis["top"] = par_new_axis(loc="bottom", axes=ax1, offset=(0, 0))
+
+    ax0.plot(self[variables[0]], self.index, "r.", label="Temperature")
+    ax1.plot(self[variables[1]], self.index, "b.", label="Salinity")
+
+    ax0.set_ylabel("Pressure [dbar]")
+    ax0.set_xlabel("Temperature [\u00b0C]")
+    ax1.set_xlabel("Salinity [kg g$^{-1}$]")
+    ax1.invert_yaxis()
+
+    try:
+        fig.suptitle(r"Station %s profile" % self.name)
+    except AttributeError:
+        pass
+
+    ax0.legend(shadow=True, fancybox=True, numpoints=1, loc="lower right")
+
+    offset = 0.01
+    x1, x2 = ax0.get_xlim()[0] - offset, ax0.get_xlim()[1] + offset
+    ax0.set_xlim(x1, x2)
+
+    offset = 0.01
+    x1, x2 = ax1.get_xlim()[0] - offset, ax1.get_xlim()[1] + offset
+    ax1.set_xlim(x1, x2)
+
+    return fig, (ax0, ax1)
+
+
+def plot_section(self, reverse=False, filled=False, **kw):
+    import gsw
+
+    lon, lat, data = list(
+        map(np.asanyarray, (self.lon, self.lat, self.values))
+    )
+    data = ma.masked_invalid(data)
+    h = self.get_maxdepth()
+    if reverse:
+        lon = lon[::-1]
+        lat = lat[::-1]
+        data = data.T[::-1].T
+        h = h[::-1]
+    lon, lat = map(np.atleast_2d, (lon, lat))
+    x = np.append(0, np.cumsum(gsw.distance(lon, lat)[0] / 1e3))
+    z = self.index.values.astype(float)
+
+    if filled:  # CAVEAT: this method cause discontinuities.
+        data = data.filled(fill_value=np.nan)
+        data = extrap_sec(data, x, z, w1=0.97, w2=0.03)
+
+    # Contour key words.
+    extend = kw.pop("extend", "both")
+    fontsize = kw.pop("fontsize", 12)
+    labelsize = kw.pop("labelsize", 11)
+    cmap = kw.pop("cmap", plt.cm.rainbow)
+    levels = kw.pop(
+        "levels",
+        np.arange(np.floor(data.min()), np.ceil(data.max()) + 0.5, 0.5),
+    )
+
+    # Colorbar key words.
+    pad = kw.pop("pad", 0.04)
+    aspect = kw.pop("aspect", 40)
+    shrink = kw.pop("shrink", 0.9)
+    fraction = kw.pop("fraction", 0.05)
+
+    # Topography mask key words.
+    dx = kw.pop("dx", 1.0)
+    kind = kw.pop("kind", "linear")
+    linewidth = kw.pop("linewidth", 1.5)
+
+    # Station symbols key words.
+    station_marker = kw.pop("station_marker", None)
+    color = kw.pop("color", "k")
+    offset = kw.pop("offset", -5)
+    alpha = kw.pop("alpha", 0.5)
+
+    # Figure.
+    figsize = kw.pop("figsize", (12, 6))
+    fig, ax = plt.subplots(figsize=figsize)
+    xm, hm = gen_topomask(h, lon, lat, dx=dx, kind=kind)
+    ax.plot(xm, hm, color="black", linewidth=linewidth, zorder=3)
+    ax.fill_between(xm, hm, y2=hm.max(), color="0.9", zorder=3)
+
+    if station_marker:
+        ax.plot(
+            x,
+            [offset] * len(h),
+            color=color,
+            marker=station_marker,
+            alpha=alpha,
+            zorder=5,
+        )
+    ax.set_xlabel("Cross-shore distance [km]", fontsize=fontsize)
+    ax.set_ylabel("Depth [m]", fontsize=fontsize)
+    ax.set_ylim(offset, hm.max())
+    ax.invert_yaxis()
+
+    ax.xaxis.set_ticks_position("top")
+    ax.xaxis.set_label_position("top")
+    ax.yaxis.set_ticks_position("left")
+    ax.yaxis.set_label_position("left")
+    ax.xaxis.set_tick_params(tickdir="out", labelsize=labelsize, pad=1)
+    ax.yaxis.set_tick_params(tickdir="out", labelsize=labelsize, pad=1)
+
+    # Color version.
+    cs = ax.contourf(
+        x,
+        z,
+        data,
+        cmap=cmap,
+        levels=levels,
+        alpha=1.0,
+        extend=extend,
+        zorder=2,
+    )  # manual=True
+    # Colorbar.
+    cb = fig.colorbar(
+        mappable=cs,
+        ax=ax,
+        orientation="vertical",
+        aspect=aspect,
+        shrink=shrink,
+        fraction=fraction,
+        pad=pad,
+    )
+    return fig, ax, cb
+
+
+def cell_thermal_mass(temperature, conductivity):
+    """
+    Sample interval is measured in seconds.
+    Temperature in degrees.
+    CTM is calculated in S/m.
+
+    """
+
+    alpha = 0.03  # Thermal anomaly amplitude.
+    beta = 1.0 / 7  # Thermal anomaly time constant (1/beta).
+
+    sample_interval = 1 / 15.0
+    a = 2 * alpha / (sample_interval * beta + 2)
+    b = 1 - (2 * a / alpha)
+    dCodT = 0.1 * (1 + 0.006 * [temperature - 20])
+    dT = np.diff(temperature)
+    ctm = -1.0 * b * conductivity + a * (dCodT) * dT  # [S/m]
+    return ctm
+
+
+def mixed_layer_depth(CT, method="half degree"):
+    if method == "half degree":
+        mask = CT[0] - CT < 0.5
+    else:
+        mask = np.zeros_like(CT)
+    return Series(mask, index=CT.index, name="MLD")
+
+
+def barrier_layer_thickness(SA, CT):
+    """
+    Compute the thickness of water separating the mixed surface layer from the
+    thermocline.  A more precise definition would be the difference between
+    mixed layer depth (MLD) calculated from temperature minus the mixed layer
+    depth calculated using density.
+
+    """
+    import gsw
+
+    sigma_theta = gsw.sigma0(SA, CT)
+    mask = mixed_layer_depth(CT)
+    mld = np.where(mask)[0][-1]
+    sig_surface = sigma_theta[0]
+    sig_bottom_mld = gsw.sigma0(SA[0], CT[mld])
+    d_sig_t = sig_surface - sig_bottom_mld
+    d_sig = sigma_theta - sig_bottom_mld
+    mask = d_sig < d_sig_t  # Barrier layer.
+    return Series(mask, index=SA.index, name="BLT")
+
+
+def derive_cnv(df):
+    """Compute SP, SA, CT, z, and GP from a cnv pre-processed cast."""
+    import gsw
+
+    cast = df.copy()
+    p = cast.index.values.astype(float)
+    cast["SP"] = gsw.SP_from_C(
+        cast["c0S/m"].values * 10.0, cast["t090C"].values, p
+    )
+    cast["SA"] = gsw.SA_from_SP(
+        cast["SP"].values, p, df._metadata["lon"], df._metadata["lat"]
+    )
+    cast["SR"] = gsw.SR_from_SP(cast["SP"].values)
+    cast["CT"] = gsw.CT_from_t(cast["SA"].values, cast["t090C"].values, p)
+    cast["z"] = -gsw.z_from_p(p, df._metadata["lat"])
+    cast["sigma0_CT"] = gsw.sigma0(cast["SA"].values, cast["CT"].values)
+    return cast
