+++ setup.cfg
@@ -9,3 +9,15 @@ versionfile_source = ctd/_version.py
 versionfile_build = ctd/_version.py
 tag_prefix = v
 parentdir_prefix =
+
+[tool:pytest]
+flake8-max-line-length = 105
+flake8-ignore =
+    E203
+    docs/* ALL
+    versioneer.py ALL
+    ctd/_version.py ALL
+
+[metadata]
+description-file = README.md
+license_file = LICENSE.txt
